foundry graders

+++ graders/github.py
@@ -17,7 +17,9 @@ def check_duplicate_repo(assignment,repo,user_id):
 def setup_repo(base_path,repo,destination):
 
 
-    subprocess.run(['git', 'clone',repo, os.path.join(base_path,destination)])
+    result = subprocess.run(['git', 'clone',repo, os.path.join(base_path,destination)])
+    if result.returncode != 0:
+        return False
     source_dirs = ['src','test','script']
     for source in source_dirs:
         shutil.rmtree(os.path.join(base_path,'foundry',source),ignore_errors=True)
@@ -26,6 +28,8 @@ def setup_repo(base_path,repo,destination):
     shutil.rmtree(os.path.join(base_path,'foundry','out'),ignore_errors=True)
     shutil.rmtree(os.path.join(base_path,'foundry','cache'),ignore_errors=True)
     shutil.rmtree(os.path.join(base_path,'foundry','broadcast'),ignore_errors=True)
+
+    return True
 
 def on_rm_error(func, path, exc_info):
     os.chmod(path, stat.S_IWRITE)  # Change the permission to writable
